breaking change: re-implement power-doctest
