--- bin/power-doctest.js
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-'use strict';
-
-require('../lib/cli').parse(process.argv.slice(2));
