--- lib/cli/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-    parse: require('./argv')
-};
