+++ package.json
@@ -7,10 +7,18 @@
     "power-doctest": "./bin/power-doctest.js"
   },
   "scripts": {
-    "test": "node_modules/mocha/bin/mocha tests"
+    "build": "babel src --out-dir lib --source-maps",
+    "watch": "babel src --out-dir lib --watch --source-maps",
+    "prepublish": "npm run --if-present build",
+    "test": "mocha"
   },
   "homepage": "https://github.com/azu/power-doctest",
-  "tags": "JavaScript,testing,document,assert,power-assert",
+  "tags": [
+    "doctest",
+    "power-assert",
+    "testing",
+    "AST"
+  ],
   "repository": {
     "type": "git",
     "url": "https://github.com/azu/power-doctest.git"
@@ -18,19 +26,27 @@
   "author": "azu",
   "license": "MIT",
   "dependencies": {
+    "ast-source": "^1.0.6",
     "bluebird": "^2.2.2",
-    "clone": "~0.1.11",
-    "escodegen": "~1.3.0",
-    "espower": "^0.8.0",
-    "esprima": "^1.2.2",
-    "estraverse": "~1.5.0",
-    "lodash": "~2.4.1",
+    "clone": "^1.0.2",
+    "comment-to-assert": "^1.0.2",
+    "empower": "^1.0.0",
+    "escodegen": "^1.6.1",
+    "espower": "^1.0.6",
+    "esprima": "^2.5.0",
+    "estraverse": "^4.1.0",
+    "lodash": "^3.10.1",
     "optimist": "~0.6.0",
-    "power-assert": "^0.8.0",
-    "rocambole": "~0.3.2"
+    "power-assert": "^0.11.0",
+    "rocambole": "^0.6.0"
   },
   "devDependencies": {
-    "chai": "~1.9.1",
-    "mocha": "^1.21.4"
+    "ast-equal": "^1.0.2",
+    "babel": "^5.8.21",
+    "chai": "^3.2.0",
+    "espower-babel": "^3.2.0",
+    "esprima": "^2.5.0",
+    "mocha": "^2.2.5",
+    "power-assert": "^0.11.0"
   }
 }
