+++ test/mocha.opts
@@ -1,0 +1,1 @@
+--compilers js:espower-babel/guess
